requirements.txt
asgiref==3.5.0
certifi==2022.6.15
charset-normalizer==2.0.12
Django==3.2.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
idna==3.3
Pillow==9.0.0
pytz==2021.3
requests==2.28.0
sqlparse==0.4.2
urllib3==1.26.9
